Fix is_exact() by changing condition precendence and add tests
matchnot in is_exact was wrongly XORed with the rest of
conditions. is_exact() is supposed to check whether a needle is found
in haystack for the given substring position and XOR that value with
matchnot.

The conditions preceeding XOR matchnot were not grouped
together previously. This resulted in the XOR operation getting
executed before all conditions before it were evaluated.